Fixing a "byte out of range" error when parsing a string to byte array Signed-off-by: Vlad Kolotov <vkolotov@gmail.com>

+++ src/main/java/org/sputnikdev/bluetooth/gattparser/BluetoothGattParser.java
@@ -325,7 +325,7 @@ public class BluetoothGattParser {
         String[] tokens = data.split(",");
         byte[] bytes = new byte[tokens.length];
         for (int i = 0; i < tokens.length; i++) {
-            bytes[i] = Byte.valueOf(tokens[i].trim(), radix);
+            bytes[i] = (byte) (int) Integer.valueOf(tokens[i].trim(), radix);
         }
         return bytes;
     }
